refactored callbacks into separate module
